Always call onBeforeDataWrite handlers, even if the session data does not change
This way, the handler may change the session data, and set additionalData too. This was documented as possible even before, but in reality setting additional data didn't really always store them to the sessions table. That's because nothing was written to the table when session data hasn't changed, no matter what was in additionalData. And setting additionalData in the handler did mostly nothing, because the handler was called only when there was a change in the session data aka the Catch-22 problem.

If anyone would need the handler to be executed only when session data has changed, this could be emulated by passing the result of `$this->data[$id] !== $data` as a param to the handler, and then adding an `if` in the handler.